Deleted unnecessary imports

+++ client/src/components/dashboard/Jobs.js
@@ -8,7 +8,9 @@ import "../../styles/jobs.css";
 
 class Jobs extends Component {
   onDelete(id) {
-    this.props.deleteJob(id);
+    if (window.confirm("Are you sure you wish to delete this item?")) {
+      this.props.deleteJob(id);
+    }
   }
 
   render() {
